cern-ci.sh: add special handling for tagged commits

+++ integration/cern-ci.sh
@@ -13,7 +13,11 @@ python -V
 localdir=/opt/home/cheby
 
 base_destdir=/acc/local/share/ht_tools/noarch/cheby
-suffix=$CI_COMMIT_SHORT_SHA
+if [ x"$CI_COMMIT_TAG" != x ]; then
+    suffix=$CI_COMMIT_TAG
+else
+    suffix=$CI_COMMIT_SHORT_SHA
+fi
 destdir=$base_destdir/cheby-$suffix
 prefix=$destdir/lib/python3.6/site-packages/
 mkdir -p $prefix
@@ -25,16 +29,18 @@ python3 ./setup.py install --prefix $destdir
 cd $base_destdir
 ln -sfn cheby-$suffix cheby-latest
 
-# Remove the old version, unless it is the same as the current one
-# (could happen when re-running CI/CD: there is no new version and
-#  the current one shouldn't be removed).
-if [ -f last ]; then
-    old=$(cat last)
-    if [ "$old" != "$suffix" ]; then
-       rm -rf ./cheby-$old
+if [ x"$CI_COMMIT_TAG" = x ]; then
+    # Remove the old version, unless it is the same as the current one
+    # (could happen when re-running CI/CD: there is no new version and
+    #  the current one shouldn't be removed).
+    if [ -f last ]; then
+        old=$(cat last)
+        if [ "$old" != "$suffix" ]; then
+            rm -rf ./cheby-$old
+        fi
     fi
+    echo $suffix > last
 fi
-echo $suffix > last
 
 #############
 # DFS update
@@ -49,6 +55,11 @@ smbclient -k //cerndfs.cern.ch/dfs/Applications/Cheby -Tx $tarfile
 
 # Remove old version
 smbclient -k //cerndfs.cern.ch/dfs/Applications/Cheby -c "rename cheby-latest cheby-old; rename cheby-$suffix cheby-latest; deltree cheby-old"
+
+if [ x"$CI_COMMIT_TAG" != x ]; then
+    # For tag commit, keep the tagged version.
+    smbclient -k //cerndfs.cern.ch/dfs/Applications/Cheby -Tx $tarfile
+fi
 
 rm -f $tarfile
